refactor(api): extract app specs and parallel assets from apidata.js
Second pass on #147. apidata.js 1369 -> 1051 lines; 1702 at the start of this
work, so 38% smaller overall.

Both blocks were fully self-contained, which is why they moved without touching
anything around them:

  api/specs.js          fetch_global_app_specs, fetch_global_app_specs_raw,
                        the homeAppSpecsRaw_v1 cache and its trimming
  api/parallelAssets.js pa_summary_full, wallet_pas_summary and their internal
                        helpers, none of which were used anywhere else

Named api/specs.js rather than api/appSpecs.js deliberately: src/appSpecs.js
already exists and owns the per-spec resource maths this calls into. Two files
with the same basename and different jobs is how an import ends up pointing at
the wrong one.

THE BUILD CAUGHT A BREAK THE TESTS DID NOT, for the third time this session,
and this one is worth recording because it is a real JS subtlety rather than
carelessness:

    export { fetch_global_app_specs_raw } from 'api/specs';

re-exports the name for IMPORTERS of apidata.js, but does NOT bind it in
apidata.js's own scope. fetch_global_stats calls it directly, so apidata.js now
carries a real import alongside the re-export. 530 tests passed throughout;
only `npx react-scripts build` reported 'fetch_global_app_specs_raw' is not
defined.

Two imports became dead once the specs block left (categorizeAppSpec,
specResources) and are removed. buildSpecIndex stays -- it is still used by
fetch_global_stats at line 577, so it was checked rather than assumed.

Behaviour-preservation shown again, not assumed: the Track 3 audit harness was
re-run against live data after the split and D1 earnings still agree with the
independent reference on all twelve figures within 1e-9.

Remaining in apidata.js: global stats, node transformation, health, GPU pricing
and DOS state. GPU/DOS was deliberately left -- it needs REQUEST_OPTIONS_API,
which five other functions in apidata.js still share, so moving it wants a
shared-constants decision rather than a duplicated literal.
